+++ src/main/java/com/example/solidconnection/auth/controller/AuthController.java
@@ -11,8 +11,8 @@ import com.example.solidconnection.auth.dto.oauth.OAuthResponse;
 import com.example.solidconnection.auth.dto.oauth.OAuthSignInResponse;
 import com.example.solidconnection.auth.service.AuthService;
 import com.example.solidconnection.auth.service.EmailSignInService;
-import com.example.solidconnection.auth.service.EmailSignUpTokenProvider;
-import com.example.solidconnection.auth.service.SignUpService;
+import com.example.solidconnection.auth.service.signup.EmailSignUpTokenProvider;
+import com.example.solidconnection.auth.service.signup.SignUpService;
 import com.example.solidconnection.auth.service.oauth.OAuthService;
 import com.example.solidconnection.common.exception.CustomException;
 import com.example.solidconnection.common.exception.ErrorCode;
