+++ src/main/java/com/example/solidconnection/auth/service/oauth/OAuthService.java
@@ -1,5 +1,6 @@
 package com.example.solidconnection.auth.service.oauth;
 
+import com.example.solidconnection.auth.domain.SignUpToken;
 import com.example.solidconnection.auth.dto.SignInResponse;
 import com.example.solidconnection.auth.dto.oauth.OAuthCodeRequest;
 import com.example.solidconnection.auth.dto.oauth.OAuthResponse;
@@ -7,7 +8,7 @@ import com.example.solidconnection.auth.dto.oauth.OAuthSignInResponse;
 import com.example.solidconnection.auth.dto.oauth.OAuthUserInfoDto;
 import com.example.solidconnection.auth.dto.oauth.SignUpPrepareResponse;
 import com.example.solidconnection.auth.service.SignInService;
-import com.example.solidconnection.auth.service.SignUpTokenProvider;
+import com.example.solidconnection.auth.service.signup.SignUpTokenProvider;
 import com.example.solidconnection.siteuser.domain.AuthType;
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import com.example.solidconnection.siteuser.repository.SiteUserRepository;
@@ -50,7 +51,7 @@ public class OAuthService {
     }
 
     private SignUpPrepareResponse getSignUpPrepareResponse(OAuthUserInfoDto userInfoDto, AuthType authType) {
-        String signUpToken = signUpTokenProvider.generateAndSaveSignUpToken(userInfoDto.getEmail(), authType);
-        return SignUpPrepareResponse.of(userInfoDto, signUpToken);
+        SignUpToken signUpToken = signUpTokenProvider.generateAndSaveSignUpToken(userInfoDto.getEmail(), authType);
+        return SignUpPrepareResponse.of(userInfoDto, signUpToken.token());
     }
 }
